+++ src/lib.rs
@@ -51,7 +51,7 @@
 //!
 //! ```toml
 //! [dependencies]
-//! rust-wistia = { version = "0.1", default-features = false, features = ["native-tls", "logging", "serde-std"] }
+//! rust-wistia = { version = "0.3", default-features = false, features = ["native-tls", "logging", "serde-std"] }
 //! ```
 //!
 //! [hyper]: https://docs.rs/hyper
@@ -61,6 +61,7 @@
 //!
 
 mod api;
+pub mod auth;
 pub mod constants;
 mod https;
 pub mod models;